Bring back non-numeric characters
Non-numeric characters were removed from numeric values, to have their
string representations correspond directly to numbers (this had to do
with adopting a CSV to HTML converter for creating HTML representation
of reports). For the time being, I will leave this as it had been previously
(not entirely happy with just generating externally the HTML reports).

For an HTML representation or a basic FE, I will probably integrate some
open source code in the future to implement it as part of the project
rather than calling a generator externally.